requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
google-generativeai==0.3.0
math3==0.0.0
geopy==2.4.1
difflib==0.0.0